+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-* Add `Dataset.x_axis_bounds` and `Dataset.y_axis_bounds` to allow setting per-dataset axis bounds.
+* Add `Dataset.x_axis_bounds` to allow setting per-dataset axis bounds.
 
 ## v0.13.0 - 2020-11-14